fix test for sum type

+++ modules/calcite/src/test/java/org/apache/ignite/internal/processors/query/calcite/planner/AggregatePlannerTest.java
@@ -321,7 +321,7 @@ public class AggregatePlannerTest extends AbstractAggregatePlannerTest {
 
         assertEquals(tf.createJavaType(Long.class), rowTypes.getFieldList().get(1).getType());
         assertEquals(tf.createJavaType(Long.class), rowTypes.getFieldList().get(2).getType());
-        assertEquals(tf.createJavaType(BigDecimal.class), rowTypes.getFieldList().get(3).getType());
+        assertEquals(tf.createJavaType(Long.class), rowTypes.getFieldList().get(3).getType());
         assertEquals(tf.createJavaType(BigDecimal.class), rowTypes.getFieldList().get(4).getType());
         assertEquals(tf.createJavaType(BigDecimal.class), rowTypes.getFieldList().get(5).getType());
         assertEquals(tf.createJavaType(Double.class), rowTypes.getFieldList().get(6).getType());
@@ -376,7 +376,7 @@ public class AggregatePlannerTest extends AbstractAggregatePlannerTest {
 
         assertEquals(tf.createJavaType(Long.class), rowTypes.getFieldList().get(1).getType());
         assertEquals(tf.createJavaType(Long.class), rowTypes.getFieldList().get(2).getType());
-        assertEquals(tf.createJavaType(BigDecimal.class), rowTypes.getFieldList().get(3).getType());
+        assertEquals(tf.createJavaType(Long.class), rowTypes.getFieldList().get(3).getType());
         assertEquals(tf.createJavaType(BigDecimal.class), rowTypes.getFieldList().get(4).getType());
         assertEquals(tf.createJavaType(BigDecimal.class), rowTypes.getFieldList().get(5).getType());
         assertEquals(tf.createJavaType(Double.class), rowTypes.getFieldList().get(6).getType());
